refactor and divide timings for vqe
