+++ .github/workflows/test.yml
@@ -69,7 +69,7 @@ jobs:
             path: docs/stats.json
 
   vqe_benchmarks:
-    name: Benchmarks VQE
+    name: Benchmarks VQE no shots
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -93,12 +93,40 @@ jobs:
         - name: Store benchmark result
           uses: actions/upload-artifact@v4
           with:
-            name: vqestats
-            path: docs/stats_vqe.json
+            name: vqestatsnoshots
+            path: docs/stats_vqe_noshots.json
+
+  vqe_benchmarks_shots:
+    name: Benchmarks VQE with shots
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.10"]
+    steps:
+        - name: Checkout main code and submodules
+          uses: actions/checkout@v4
+        - name: Set up Python ${{ matrix.python-version }}
+          uses: actions/setup-python@v5
+          with:
+            python-version: ${{ matrix.python-version }}
+        - name: Install hatch
+          run: |
+            python -m pip install hatch
+        - name: Install main dependencies
+          run: |
+            python -m hatch -v -e tests
+        - name: Perform benchmarks
+          run: |
+            python -m hatch -e tests run vqeshots
+        - name: Store benchmark result
+          uses: actions/upload-artifact@v4
+          with:
+            name: vqestatsshots
+            path: docs/stats_vqe_shots.json
 
   test_docs:
     name: Documentation
-    needs: [test_benchmarks, vqe_benchmarks]
+    needs: [test_benchmarks, vqe_benchmarks, vqe_benchmarks_shots]
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -124,7 +152,11 @@ jobs:
         - name: Download benchmark VQE result
           uses: actions/download-artifact@v4
           with:
-            name: vqestats
+            name: vqestatsnoshots
+        - name: Download benchmark VQE result shots
+          uses: actions/download-artifact@v4
+          with:
+            name: vqestatsshots
         - name: Test docs
           run: |
             python -m hatch -e docs run mkdocs build --clean --strict
