+++ aquordion/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import annotations
 
-from .benchmarks import expectation_horqrux, expectation_pyq, run_horqrux, run_pyq
+from .api_benchmarks import expectation_horqrux, expectation_pyq, run_horqrux, run_pyq
 from .utils import values_to_jnp
