+++ docs/benchmarks.md
@@ -87,19 +87,34 @@ print(fig_to_html(plt.gcf())) # markdown-exec: hide
 Here are the median execution times for VQE. We compare optimizing with `PyQTorch` against optimizing with `Horqrux` and jitting.
 
 ```python exec="on" source="material-block" session="benchmarks"
-fname = "stats_vqe.json"
+fname = "stats_vqe_noshots.json"
+fnameshots = "stats_vqe_shots.json"
+
 if not os.path.isfile(fname):
-    fname = "docs/stats_vqe.json"
+    fname = "docs/stats_vqe_noshots.json"
+    fnameshots = "docs/stats_vqe_shots.json"
 with open(fname, 'r') as f:
     data_vqe = json.load(f)['benchmarks']
+with open(fnameshots, 'r') as f:
+    data_vqeshots = json.load(f)['benchmarks']
 
 data_stats_vqe = [{'name': x['name']} | x['params'] | x['stats'] for x in data_vqe]
+data_stats_vqeshots = [{'name': x['name']} | x['params'] | x['stats'] for x in data_vqeshots]
 
 frame_vqe = pd.DataFrame(data_stats_vqe)
 frame_vqe['name'] = frame_vqe['name'].apply(lambda x: re.findall('test_(.*)\\[', x)[0])
 frame_vqe['fn_circuit'] = frame_vqe['benchmark_vqe_ansatz'].apply(str)
 frame_vqe['fn_circuit'] = frame_vqe['fn_circuit'].apply(lambda x: re.findall('function (.*) at', x)[0])
 frame_vqe['name'] = frame_vqe['name'].str.replace('vqe_', '')
+frame_vqe['n_shots'] = 0
+
+frame_vqeshots = pd.DataFrame(data_stats_vqeshots)
+frame_vqeshots['name'] = frame_vqeshots['name'].apply(lambda x: re.findall('test_(.*)\\[', x)[0])
+frame_vqeshots['fn_circuit'] = frame_vqeshots['benchmark_vqe_ansatz'].apply(str)
+frame_vqeshots['fn_circuit'] = frame_vqeshots['fn_circuit'].apply(lambda x: re.findall('function (.*) at', x)[0])
+frame_vqeshots['name'] = frame_vqeshots['name'].str.replace('vqe_', '')
+frame_vqeshots['n_shots'] = 1000
+frame_vqeshots['diff_mode'] = 'ad'
 ```
 
 ### Times
@@ -107,7 +122,7 @@ frame_vqe['name'] = frame_vqe['name'].str.replace('vqe_', '')
 Below we present the distribution of median times for each circuit type, without shots and with shots ($n_shots = 1000$).
 
 ```python exec="on" source="material-block" session="benchmarks"
-axes = frame_vqe[frame_vqe.n_shots == 0].boxplot('median', by=['fn_circuit', 'name', 'diff_mode'])
+axes = frame_vqe.boxplot('median', by=['fn_circuit', 'name', 'diff_mode'])
 axes.set_title('Timing distributions by test and circuit without shots - 50 epochs')
 axes.set_xlabel('')
 axes.set_ylabel('Time (s)')
@@ -119,7 +134,7 @@ print(fig_to_html(plt.gcf())) # markdown-exec: hide
 
 
 
-axes = frame_vqe[frame_vqe.n_shots > 0].boxplot('median', by=['fn_circuit', 'name', 'diff_mode'])
+axes = frame_vqeshots.boxplot('median', by=['fn_circuit', 'name'])
 axes.set_title('Timing distributions by test and circuit with shots  - 10 epochs')
 axes.set_xlabel('')
 axes.set_ylabel('Time (s)')
@@ -136,6 +151,8 @@ print(fig_to_html(plt.gcf())) # markdown-exec: hide
 Below we present the distribution of median speed-ups for each circuit type. The timing ratio between `PyQTorch` (numerator) and `Horqrux` (denominator) executions is computed. A ratio higher than $1$ means `Horqrux` and jitting provide computational speed-up over `PyQTorch`.
 
 ```python exec="on" source="material-block" session="benchmarks"
+
+frame_vqe = pd.concat([frame_vqe, frame_vqeshots], ignore_index=True)
 
 pyq_vqe = frame_vqe[frame_vqe.name == 'pyq'][['fn_circuit', 'median', 'n_shots', 'diff_mode']]
 horqrux_vqe = frame_vqe[frame_vqe.name == 'horqrux'][['fn_circuit', 'median', 'n_shots', 'diff_mode']]
