+++ pyproject.toml
@@ -42,9 +42,11 @@ features = [
 ]
 
 [tool.hatch.envs.tests.scripts]
-test = "pytest -n auto --ignore=tests/test_timings.py --ignore=tests/test_timings_vqe.py {args}"
-benchmarks = "pytest --benchmark-disable-gc --benchmark-json docs/stats.json -v -s tests/test_timings.py"
-vqe = "pytest --benchmark-disable-gc --benchmark-json docs/stats_vqe.json -v -s tests/test_timings_vqe.py"
+test = "pytest -n auto --ignore=tests/timings/ {args}"
+benchmarks = "pytest --benchmark-disable-gc --benchmark-json docs/stats.json -v -s tests/timings/test_timings.py"
+vqe = "pytest --benchmark-disable-gc --benchmark-json docs/stats_vqe_noshots.json -v -s tests/timings/vqe/test_noshots.py"
+vqeshots = "pytest --benchmark-disable-gc --benchmark-json docs/stats_vqe_shots.json -v -s tests/timings/vqe/test_shots.py"
+
 
 [tool.hatch.envs.docs]
 installer = "uv"
