+++ tests/test_benchmarks_circuits.py
@@ -9,7 +9,7 @@ from qadence import AbstractBlock, QuantumCircuit
 from qadence.backends.jax_utils import jarr_to_tensor
 from qadence.states import equivalent_state
 
-from aquordion.benchmarks import expectation_horqrux, expectation_pyq, run_horqrux, run_pyq
+from aquordion.api_benchmarks import expectation_horqrux, expectation_pyq, run_horqrux, run_pyq
 from aquordion.utils import values_to_jnp
 
 ATOL_32 = 1e-07  # 32 bit precision
